made changes removed setinterval and made everything smoother

+++ src/js/enemy.js
@@ -1,43 +1,28 @@
 import { Actor, Vector } from 'excalibur';
-import { Resources } from './resources.js';
+import { Resources } from './resources';
+import { Player } from './player.js';
 
 export class Enemy extends Actor {
     isDead = false;
 
-    constructor(x, y, player) {
+    constructor(x, y) {
         super({ width: 40, height: 40 });
         this.pos = new Vector(x, y);
-        this.player = player;
         this.graphics.use(Resources.Fish.toSprite());
-    }
-
-    onPreUpdate(engine, delta) {
-        this.vel = this.isDead ? Vector.Zero : new Vector(-100, 0);
+        this.vel = new Vector(-100, 0); // Move left across the screen
     }
 
     onInitialize(engine) {
         this.on('collisionstart', (event) => {
             const other = event.other?.owner;
-            // Only affect Player class
-            if (other instanceof import('./player.js').Player) {
-                if (other.game && typeof other.game.loseLife === 'function') {
-                    other.game.loseLife(1);
-                    // Game.js loseLife updates UI
-                    if (other.game.lives.get() <= 0) {
-                        setTimeout(() => {
-                            window.location.reload();
-                        }, 500);
-                    } else {
-                        // Respawn player at starting position
-                        other.pos = new Vector(400, 400);
-                        other.vel = Vector.Zero;
-                    }
-                }
+            if (other instanceof Player && other.game && typeof other.game.loseLife === 'function') {
+                other.game.loseLife(1);
             }
         });
     }
 
     die() {
+        if (this.isDead) return;
         this.isDead = true;
         this.graphics.use(Resources.Bones.toSprite());
         setTimeout(() => this.kill(), 500);
